of/dynamic: Update list of aliases on aliases changes
Currently the list of aliases is not updated when an overlay that
modifies /aliases is added or removed. This breaks drivers (e.g. serial)
that rely on of_alias_get_id().

Update the list of aliases when a property of the /aliases node is
added, removed, or updated.

FIXME Handle this in the OF core itself, by registering a notifier using
      of_reconfig_notifier_register()
FIXME Handle adding/removing the "aliases" node itself.

Signed-off-by: Geert Uytterhoeven <geert+renesas@glider.be>