of: fix the error message handling
The previous patch to add back an error message didn't take the
successful case into account. This results in the following trace
showing when a symbol is properly resolved:
OF: resolver: overlay phandle fixup failed: 0

So remove the 'err_out' label and test 'err' value in the 'out' label
before printing the error message.

Signed-off-by: Gary Bisson <gary.bisson@boundarydevices.com>